reyyandemir_calismalari/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic[email]==2.10.3
python-multipart==0.0.19
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
Pillow==11.0.0
numpy==2.2.0
face-recognition==1.3.0
opencv-python==4.10.0.84
